+++ examples/generate.py
@@ -6,13 +6,12 @@ import sys
 from typing import List
 
 import torch
-
 from curated_transformers.generation.config import (
     GreedyGeneratorConfig,
     SampleGeneratorConfig,
 )
 from curated_transformers.layers.attention import enable_torch_sdp
-from curated_transformers.util.auto_model import AutoGenerator
+from curated_transformers.models.auto_model import AutoGenerator
 
 EPILOG = """This program takes a JSON list of strings from the standard inputs as prompts. For example:
